bugs/agents/requirements.txt
langchain-openai>=0.0.5
pydantic>=2.0.0
python-dotenv>=1.0.0
playwright>=1.49.0
browser-use>=0.1.37 